+++ NEWS.md
@@ -24,4 +24,7 @@
 * Implemented temporary solution to rbind and bind_rows using vctrs
 * Defined a method for `dplyr::slice`
 * Added tests
+* Implemented `rt_bind_rows` and `rt_add_row` as robust and efficient counterparts
+ of `dplyr::bind_rows` and `tibble::add_row`, which we can unfortunately not use
+ on reactibbles reliably.